refactor sql tests
